to create the Cart in DOM
