+++ README.md
@@ -6,9 +6,9 @@
 - To use **BEM** notation;
 - The data on the selected items for purchase can be located in **'localStorage'**;
 - The purchase registration takes three stages, which separately show the inputs;
-- To realize the search of the countries from conditionally fetched data.
+- To realize the search of the countries from (conditionally) fetched data.
 When filling the input of the Country - to search and show the matching;
-- To write all class-names, etc.(connected to DOM) at the head of the script for
+- To write all class-names, etc.(connected to DOM) in the separate file for
 the convenience in possible edit;
 ### Stack: Gulp4, SCSS, ES-2015;
