+++ src/index.html
@@ -46,8 +46,8 @@
                 <h3>Recipient</h3>
                 <div class="input-wrapper">
                     <input type="text" name="recipient-full-name"
-                           class="form__input" placeholder="Full Name"
-                           title="" maxlength="30" required
+                           class="form__input" placeholder="Name Surname"
+                           title="Jonathan Smith" maxlength="30" required
                     >
                 </div>
                 <div class="flex-box left">
@@ -55,7 +55,7 @@
                         <input type="text" name="recipient-phone-number"
                                class="form__input part-width-smaller"
                                placeholder="Daytime Phone"
-                               title="" maxlength="20" required
+                               title="+7(987)123-77-88" maxlength="20" required
                         >
                     </div>
                     <div class="form__block__comment">
@@ -66,20 +66,20 @@
                 <div class="input-wrapper">
                     <input type="text" name="recipient-street-address"
                            class="form__input" placeholder="Street Address"
-                           title="" maxlength="25" required
+                           title="Pushkina" maxlength="25" required
                     >
                 </div>
                 <div class="input-wrapper">
                     <input type="text" name="optional"
                            class="form__input"
                            placeholder="Apt, Suite, Bldg, Gate Code. (optional)"
-                           title="" maxlength="25"
+                           title="format: All except: '+', '=', '*'" maxlength="25"
                     >
                 </div>
                 <div class="input-wrapper target-sign push-top_30">
                     <input type="text" name="recipient-city"
                            class="form__input" placeholder="City"
-                           title="" maxlength="25" required
+                           title="Kyiv" maxlength="25" required
                     >
                 </div>
                 <div class="flex-box between">
@@ -91,6 +91,7 @@
                             <input type="text" name="recipient-country"
                                    class="form__input part-width-smaller"
                                    placeholder="Country"
+                                   title="please, choose the Country"
                                    maxlength="25" required
                             >
                         </div>
@@ -99,6 +100,7 @@
                         <input type="text" name="recipient-zip"
                                class="form__input part-width-small"
                                placeholder="ZIP"
+                               title="420015"
                                maxlength="8" required
                         >
                     </div>
@@ -113,33 +115,34 @@
                 <div class="input-wrapper">
                     <input type="text" name="billing-full-name"
                            class="form__input" placeholder="Full Name"
-                           title="" maxlength="30" required
+                           title="Jonathan Smith" maxlength="30" required
                     >
                 </div>
                 <div class="input-wrapper">
                     <input type="email" name="billing-email"
                            class="form__input" placeholder="Email Address"
-                           title="" maxlength="30" required
+                           title="j.smith@gmail.com" maxlength="30" required
                     >
                 </div>
                 <h3>Billing Address</h3>
                 <div class="input-wrapper">
                     <input type="text" name="billing-street-address"
                            class="form__input" placeholder="Street Address"
-                           title="" maxlength="25" required
+                           title="Pushkina" maxlength="25" required
                     >
                 </div>
                 <div class="input-wrapper">
                     <input type="text" name="billing-optional"
                            class="form__input"
                            placeholder="Apt, Suite, Bldg, Gate Code. (optional)"
+                           title="format: All except: '+', '=', '*'"
                            maxlength="25" title=""
                     >
                 </div>
                 <div class="input-wrapper target-sign push-top_30">
                     <input type="text" name="billing-city"
                            class="form__input" placeholder="City"
-                           title="" maxlength="25" required
+                           title="Kyiv" maxlength="25" required
                     >
                 </div>
                 <div class="flex-box between">
@@ -151,6 +154,7 @@
                             <input type="text" name="billing-country"
                                    class="form__input part-width-smaller"
                                    placeholder="Country"
+                                   title="Please, choose the Country"
                                    maxlength="25" required
                             >
                         </div>
@@ -159,6 +163,7 @@
                         <input type="text" name="billing-zip"
                                class="form__input part-width-small"
                                placeholder="ZIP"
+                               title="420015"
                                maxlength="8" required
                         >
                     </div>
@@ -177,7 +182,7 @@
                     <input type="text" name="card-name"
                            class="form__input"
                            placeholder="Name as it appears on your card"
-                           title="" maxlength="30" required
+                           title="Jonathan Smith" maxlength="30" required
                     >
                 </div>
                 <h3>Card Number</h3>
@@ -185,6 +190,7 @@
                     <input type="text" name="card-number"
                            class="form__input"
                            placeholder="XXXX XXXX XXXX XXXX XXXX"
+                           title="Please, choose the card number 20 digits"
                            maxlength="24" title="" required
                     >
                 </div>
@@ -195,6 +201,7 @@
                             <input type="text" name="card-date"
                                    class="form__input part-width-smallest"
                                    placeholder="MM / YY" title=""
+                                   title="MMYY: 0518"
                                    maxlength="7" required
                             >
                         </div>
@@ -204,7 +211,7 @@
                         <div class="input-wrapper">
                             <input type="text" name="card-code"
                                    class="form__input part-width-smallest"
-                                   title="" maxlength="4" required
+                                   title="0458" maxlength="4" required
                             >
                         </div>
                     </div>
@@ -262,7 +269,6 @@
         <div>
             <h3>Vadim Komolov</h3>
             <p>e-mail: vkomolov1974@gmail.com</p>
-            <p>telephone: +(380) 66 335-28-98</p>
         </div>
     </div>
 </footer>
